+++ crates/lib/src/compiler/quilc.rs
@@ -228,7 +228,7 @@ pub enum CompilationError {
     Libquil(crate::compiler::libquil::Error),
     /// Errors during compilation when using RPCQ
     #[error("compilation error from RPCQ: {0}")]
-    Rpcq(crate::compiler::rpcq::Error),
+    Rpcq(rpcq::Error),
 }
 
 /// The response from quilc for a `quil_to_native_quil` request.
@@ -358,7 +358,7 @@ mod tests {
     }
 
     async fn rpcq_client() -> rpcq::Client {
-        let qcs = Qcs::load().await;
+        let qcs = Qcs::load();
         let endpoint = qcs.get_config().quilc_url();
         rpcq::Client::new(endpoint).unwrap()
     }
@@ -387,7 +387,7 @@ MEASURE 1 ro[1]
 
     #[tokio::test]
     async fn run_compiled_bell_state_on_qvm() {
-        let client = Qcs::load().await;
+        let client = Qcs::load();
         let client = qvm::http::HttpClient::from(&client);
         let output = rpcq_client()
             .await
@@ -398,7 +398,7 @@ MEASURE 1 ro[1]
                 CompilerOpts::default(),
             )
             .expect("Could not compile");
-        let mut results = crate::qvm::Execution::new(&output.program.to_quil_or_debug())
+        let mut results = qvm::Execution::new(&output.program.to_quil_or_debug())
             .unwrap()
             .run(
                 NonZeroU16::new(10).expect("value is non-zero"),