+++ crates/lib/src/qpu/translation.rs
@@ -13,6 +13,8 @@ use qcs_api_client_grpc::{
     },
 };
 use tokio::time::error::Elapsed;
+#[cfg(feature = "tracing")]
+use tracing::instrument;
 
 use crate::client::{GrpcClientError, Qcs, DEFAULT_HTTP_API_TIMEOUT};
 
@@ -42,6 +44,7 @@ pub struct EncryptedTranslationResult {
 }
 
 /// Translate a program, returning an encrypted and translated program.
+#[cfg_attr(feature = "tracing", instrument(skip_all))]
 pub async fn translate<TO>(
     quantum_processor_id: &str,
     quil_program: &str,