+++ crates/lib/tests/basic_qvm.rs
@@ -17,13 +17,13 @@ MEASURE 1 second
 "##;
 
 async fn quilc_client() -> rpcq::Client {
-    let qcs = Qcs::load().await;
+    let qcs = Qcs::load();
     let endpoint = qcs.get_config().quilc_url();
     rpcq::Client::new(endpoint).unwrap()
 }
 
 async fn qvm_client() -> qvm::http::HttpClient {
-    let qcs = Qcs::load().await;
+    let qcs = Qcs::load();
     qvm::http::HttpClient::from(&qcs)
 }
 
@@ -33,7 +33,7 @@ async fn test_bell_state() {
 
     let data = Executable::from_quil(PROGRAM)
         .with_quilc_client(Some(quilc_client().await))
-        .with_qcs_client(Qcs::load().await)
+        .with_qcs_client(Qcs::load())
         .with_shots(shots)
         .read_from("first")
         .read_from("second")