+++ crates/lib/tests/mocked_qpu.rs
@@ -64,7 +64,7 @@ async fn setup() {
 }
 
 async fn quilc_client() -> rpcq::Client {
-    let qcs = Qcs::load().await;
+    let qcs = Qcs::load();
     let endpoint = qcs.get_config().quilc_url();
     rpcq::Client::new(endpoint).unwrap()
 }